feat: refactor loggers to be more specific about whether they should be used client side vs server side
