+++ .github/workflows/cypress-mocked.yml
@@ -40,7 +40,8 @@ jobs:
           LOGGING_KINESIS_STREAM: ''
           EC2_INSTANCE_ID: ''
           # This is a public URL+key sent to Sentry on the client side
-          SENTRY_DSN: ${{ secrets.SENTRY_DSN }}
+          # We disable Sentry for our tests
+          SENTRY_DSN: ''
         with:
           build: yarn build
           start: yarn start, yarn mock-server
