+++ src/server/lib/idapi/verifyEmail.ts
@@ -7,7 +7,7 @@ import {
 } from '@/server/lib/IDAPIFetch';
 
 import { IdapiErrorMessages, VerifyEmailErrors } from '@/shared/model/Errors';
-import { logger } from '@/server/lib/winstonLogger';
+import { logger } from '@/server/lib/serverSideLogger';
 import { IdapiError } from '@/server/models/Error';
 
 const handleError = ({ error, status = 500 }: IDAPIError) => {
