+++ src/server/lib/middleware/logger.ts
@@ -1,5 +1,5 @@
 import { NextFunction, Request, Response } from 'express';
-import { logger } from '@/server/lib/winstonLogger';
+import { logger } from '@/server/lib/serverSideLogger';
 
 export const loggerMiddleware = (
   req: Request,
